Provide a way to create a new grid from an existing grid
Every 'tick', the grid will need to be updated. Calling this method with a
grid will transform each item (cell) in that grid with the provided block.

My intention is that I can pass a grid to it and the block will contain the
logic the decides whether or not the cell survives.