Wrap states in a Module instead of a Class
The `State` class was essentially only wrapping a couple of constants. Apart
from the random method, there's no logic here and no real need for it to be a
class.

The name change reflects that the module is a wrapper for multiple states
rather than a class than represents a single state (which it never was).