Add Neighbours class
Given coordinates and a grid, this class works out which objects neighbour a
the specified coordinates. Knowing which cells neighbour another cell is
essential when decicding whether a cell lives or dies.

`#local_group` is a bit clunky but I'm not sure how to refactor it at the
moment. I've used `Array#rotate` so that if the target cell is at the edge of
the grid, I can get its neighbouring cells from the other side of the grid
(the grid wraps, like the levels on Pacman – Pacman leaves one side of the
screen and enters on the other (on a row that is four cells long, the fourth
column neighbours the first column).