Return a new cell with a state determined by the states of its neighbours
This is the logic of the game. If I pass the constructor a cell and an array
of neighbours, it will return a new cell with the calculated state.

There's a lot of conditionals in this method but I'm okay with that as it's
essentially a factory method.

Also, it's probably primitive obsession, having an array as an argument but
I'm not convinced that `Neighbours` should care about the state of the cells
it represents.